+++ src/constrollers/wordController.js
@@ -1,4 +1,5 @@
 const wordModel = require('../models/wordModel');
+const {Op} = require("sequelize");
 
 
 /**
@@ -37,10 +38,14 @@ getWordById = async (req, res) => {
     #swagger.tags = ['Words']
     #swagger.description = 'Récupère un mot par son id.'
     #swagger.responses[200] = {description : 'Mot récupéré.'}
+    #swagger.responses[404] = {description : 'Aucun mot trouvé avec cet id.'}
     #swagger.responses[500] = {description : 'Erreur survenue lors de la récuperation d\'un mot.'}
     #swagger.parameters['id'] = {description : 'Id du mot à rechercher.', required : true}
     */
     try {
+        if (!await wordModel.findByPk(req.params.id)) {
+            return res.status(404).send("Aucun mot trouvé avec cet id.");
+        }
         const data = await wordModel.findByPk(req.params.id)
         res.status(200).send(data);
     } catch (error) {
@@ -62,6 +67,7 @@ addWord = async (req, res) => {
     #swagger.tags = ['Words']
     #swagger.description = 'Ajoute un mot.'
     #swagger.responses[201] = {description : 'Ajout effectué'}
+    #swagger.responses[400] = {description : 'Ce mot existe déjà.'}
     #swagger.responses[500] = {description : 'Erreur survenue lors de l\'ajout d\'un mot.'}
     #swagger.parameters['Mot à ajouter.'] = {
         in: 'body',
@@ -91,6 +97,9 @@ addWord = async (req, res) => {
     }
     */
     try {
+        if (await wordModel.findOne({where: {word: req.body.word}})) {
+            return res.status(400).send("Ce mot existe déjà.");
+        }
         await wordModel.create(req.body)
         res.status(201).send("Ajout effectué");
     } catch (error) {
@@ -112,6 +121,8 @@ modifyWord = async (req, res) => {
     #swagger.tags = ['Words']
     #swagger.description = 'Modifie un mot'
     #swagger.responses[200] = {description : 'Modification effectuée'}
+    #swagger.responses[400] = {description : 'Ce mot existe déjà.'}
+    #swagger.responses[404] = {description : 'Aucun mot trouvé avec cet id.'}
     #swagger.responses[500] = {description : 'Erreur survenue lors de la modification d\'un mot.'}
     #swagger.parameters['id'] = {description : 'Id du mot à modifier.', required : true}
     #swagger.parameters['Champs à modifier'] = {
@@ -142,6 +153,14 @@ modifyWord = async (req, res) => {
     }
      */
     try {
+        if (!await wordModel.findByPk(req.params.id)) {
+            return res.status(404).send("Aucun mot trouvé avec cet id.");
+        }
+
+        if (await wordModel.findOne({where: {word: req.body.word}, id: {[Op.not]: req.params.id}})) {
+            return res.status(400).send("Ce mot existe déjà.");
+        }
+
         await wordModel.update(req.body, {
             where: {
                 id: req.params.id
@@ -167,10 +186,14 @@ deleteWord = async (req, res) => {
     #swagger.tags = ['Words']
     #swagger.description = 'Supprime un mot.'
     #swagger.responses[200] = {description : 'Suppression effectuée.'}
+    #swagger.responses[404] = {description : 'Aucun mot trouvé avec cet id.'}
     #swagger.responses[500] = {description : 'Erreur survenue lors de la suppression d\'un mot.'}
     #swagger.parameters['id'] = {description : 'Id du mot.', required : true}
     */
     try {
+        if (!await wordModel.findByPk(req.params.id)) {
+            return res.status(404).send("Aucun mot trouvé avec cet id.");
+        }
         await wordModel.destroy({
             where: {
                 id: req.params.id
